pyproject.toml
[project]
name = "slope-check-backend"
version = "0.1.0"
description = "A Python package for slope checking"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.126.0",
    "pydantic==2.12.5",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.34.0",
    "beautifulsoup4>=4.12.0",
]
authors = [
    { name = "Felix Jost", email = "fel.jost@gmail.com" }
]


[project.scripts]
slope-check-backend = "slope_check_backend:main"

[project.urls]
Homepage = "https://github.com/feljost/slope-check-backend"
Repository = "https://github.com/feljost/slope-check-backend"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/slope_check_backend"]
